% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_columns}
\alias{check_columns}
\title{simple wrapper to check if important columns are missing}
\usage{
check_columns(df, columns)
}
\description{
simple wrapper to check if important columns are missing
}
\keyword{internal}
