% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composite.R
\name{composite}
\alias{composite}
\title{Composite Indicator (CI) of corruption risk in emergencies Synthetic indicator of corruption risk by contracting authority, company province and municipality}
\usage{
composite()
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\section{Motivation (deliverable 3.3)}{

Normalisation is required to make the indicators comparable. It is necessary when they have different measurement units, or a particularly skew distribution or different polarity (i.e., positive and negative)
Weighting reflects the relative importance of the individual indicators. It is one of the most influential phases when developing a CI and three main alternatives are possible: a) no weighting, implying that equal weights are applied to all individual indicators; b) subjective or expert weighting, where group of specialists define weights for each indicator; and c) objective or ‘data-driven’ weighting
Aggregation is needed for summarising individual indicator values into a single number (i.e. the composite indicator). Techniques to summarise indicators can be compensatory, partially compensatory or non-compensatory.
}

\section{Motivation (deliverable 3.4)}{

Individual indicators are often chosen paying little attention to the relationships among them, leading to the “indicator rich but information poor” environment (i.e., composite indicators that do not contribute to add knowledge).
Multivariate analyses are then needed for the study of the relational or dimensional structure of the data, to gauge the suitability of the data at issue for representing the latent phenomenon under study and to appreciate the implications of other methodological choices taken during the process of CI construction.
}

\section{Motivation (deliverable 3.5)}{

Sensitivity analysis is necessary to assess how the variation in the output (CI values) depends upon the methodological choices adopted during the process of CI development. In fact, such choices - i.e., the selection of individual indicators, the choice of the normalisation method, of the weighting system, and of the aggregation scheme - are all potential sources of uncertainty that affect the values assumed by the CI.
}

\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
