% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composite.R
\name{composite_sensitivity_methods}
\alias{composite_sensitivity_methods}
\title{composite_sensitivity_methods}
\usage{
composite_sensitivity_methods(
  indicator_list,
  cutoff = c(0.9, 0.95, 0.99, 0.995),
  ...
)
}
\description{
\code{composite_sensitivity_methods} performs the sensitivity analysis about
\strong{only the methodological choices} for computing the composite: \emph{normalisation} (cut-off values),
method for \emph{managing missing values} and \emph{weighting} scheme.
\strong{It is an internal function.}
}
\details{
Perform sensitivity analysis (only about methods) of a composite indicator
}
\keyword{internal}
