% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-winning-rate-across-crisis.R
\name{compute_barnard}
\alias{compute_barnard}
\title{compute Barnard test https://en.wikipedia.org/wiki/Barnard\%27s_test}
\usage{
compute_barnard(a, b, c, d, method = "boschloo")
}
\description{
compute Barnard test pvalue and estimate in piped expression
}
\keyword{internal}
