% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composite.R
\name{compute_composite}
\alias{compute_composite}
\title{Compute composite indicator}
\usage{
compute_composite(
  indicator_list,
  norm_method = "binary",
  miss_method = 1,
  weight_method = "experts",
  aggr_method = "linear",
  cutoff = 0.95,
  expert_weights = NULL,
  ...
)
}
\arguments{
\item{indicator_list}{list of outputs about each indicator computable for the target unit
(e.g., company or contracting authority), as returned by \code{\link[=ind_all]{ind_all()}}.}

\item{norm_method}{normalisation method (see \code{\link[=normalise]{normalise()}}).}

\item{miss_method}{missing management method (see \code{\link[=manage_missing]{manage_missing()}}).}

\item{weight_method}{weighting method (see \code{\link[=get_weights]{get_weights()}}).}

\item{aggr_method}{aggregation method (see \code{\link[=aggregate]{aggregate()}}).}

\item{cutoff}{threshold for dichotomising the indicators (when \code{norm_method = "binary"}).}

\item{expert_weights}{mean weights over a pool of anticorruption experts which express the relative impact of each indicator over the composite (these are originally expressed on a scale of 1 to 10)}

\item{...}{optional arguments for \code{\link[mirt:mirt]{mirt::mirt()}} function. See Details.}
}
\value{
indicator schema as from \code{\link[=generate_indicator_schema]{generate_indicator_schema()}}
}
\description{
\code{compute_composite} is a generic function that calculates the composite indicators according to
specified \emph{normalisation}, \emph{missing management}, \emph{weighting} and \emph{aggregation} methods.
}
\examples{
\dontrun{
if (interactive()) {
  # sample of 100k contracts
  set.seed(12345)
  i <- sample(1:nrow(mock_data_core), size = 1e5)
  mock_sample0 <- mock_data_core[sort(i), ]

  # indicators for companies
  mock_sample <- tidyr::unnest(mock_sample0, aggiudicatari, keep_empty = TRUE)
  mock_sample_variants <- tidyr::unnest(mock_sample, varianti, keep_empty = TRUE)

  out_companies <- ind_all(
    data = mock_sample,
    data_ind8 = mock_sample_variants,
    emergency_name = "coronavirus",
    target_unit = "companies"
  )
  composite_companies <- compute_composite(
    indicator_list = out_companies,
    norm_method = "binary",
    miss_method = 0,
    cutoff = 0.99,
    weight_method = "experts",
    aggr_method = "linear"
  )
}
}
}
