% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_tests.R
\name{compute_fisher}
\alias{compute_fisher}
\title{compute Fisher-exact test https://en.wikipedia.org/wiki/Fisher\%27s_exact_test}
\usage{
compute_fisher(a, b, c, d)
}
\description{
compute fisher test pvalue and estimate in piped expression
}
\keyword{internal}
