% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_tests.R
\name{compute_kolmogorov_smirnoff}
\alias{compute_kolmogorov_smirnoff}
\title{compute Kolmogorov Smirnov test in dplyr https://it.wikipedia.org/wiki/Test_di_Kolmogorov-Smirnov}
\usage{
compute_kolmogorov_smirnoff(data, var, group, alternative = "less")
}
\description{
compute Kolmogorov Smirnov test pvalue
}
\keyword{internal}
