% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_tests.R
\name{compute_prop_test}
\alias{compute_prop_test}
\title{compute Z-test proportional}
\usage{
compute_prop_test(a, b, c, d, correct = FALSE)
}
\description{
compute Z-test pvalue and estimate in piped expression
}
\keyword{internal}
