% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-lengthy-contract.R
\name{compute_ttest}
\alias{compute_ttest}
\title{compute t test}
\usage{
compute_ttest(mean_to_compare, ground_mean)
}
\description{
compute  test pvalue and estimate in piped expression
}
\keyword{internal}
