% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_tests.R
\name{compute_unpaired_ttest}
\alias{compute_unpaired_ttest}
\title{compute unpaired t-test test}
\usage{
compute_unpaired_ttest(.data, var, group, alternative = "less", paired = FALSE)
}
\description{
compute unpaired t test
}
\keyword{internal}
