% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_tests.R
\name{compute_wilcox}
\alias{compute_wilcox}
\title{compute Wilcoxon-Mann-Whitney test in dplyr https://it.wikipedia.org/wiki/Test_di_Wilcoxon-Mann-Whitney}
\usage{
compute_wilcox(
  data,
  var,
  group,
  exact = TRUE,
  alternative = "greater",
  paired = FALSE
)
}
\description{
compute Wilcoxon-Mann-Whitney test pvalue
}
\keyword{internal}
