% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-awd-eco-value-across-crisis.R
\name{compute_wilcox_test}
\alias{compute_wilcox_test}
\title{compute Wilcox test in dplyr}
\usage{
compute_wilcox_test(data, var, group, exact = TRUE, alternative = "less")
}
\description{
compute Wilcox test pvalue
}
\keyword{internal}
