% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coresoi-package.R
\docType{package}
\name{coresoi-package}
\alias{coresoi}
\alias{coresoi-package}
\title{coresoi: Measuring Corruption Risk in Emergency}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A package to calculate CORE red flags and to play with our data
}
\seealso{
Useful links:
\itemize{
  \item \url{https://core-forge.github.io/coresoi/}
  \item \url{https://github.com/CORE-forge/coresoi}
}

}
\author{
\strong{Maintainer}: Niccolo Salvini \email{niccolo.salvini27@gmail.com} (\href{https://orcid.org/0000-0002-3444-4094}{ORCID})

Authors:
\itemize{
  \item Simone DelSarto \email{simone.delsarto@unipg.it} (\href{https://orcid.org/0000-0003-1102-2881}{ORCID})
  \item Michela Gnaldi \email{michela.gnaldi@unipg.it} (\href{https://orcid.org/0000-0002-2785-3279}{ORCID})
}

}
\keyword{internal}
