% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{download_and_unzip}
\alias{download_and_unzip}
\title{Download and extract zip from target url}
\usage{
download_and_unzip(zip_url)
}
\description{
Download and extract zip from target url
}
\keyword{internal}
