% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{emergency_dates}
\alias{emergency_dates}
\title{get the emergency date from input string}
\usage{
emergency_dates(emergency_name)
}
\arguments{
\item{emergency_name}{emergency name character string}
}
\value{
a lubridate class dmy date.
}
\description{
gets the outbreak starting date from emergency name input string
}
\examples{
\dontrun{
if (interactive()) {
  emergency_dates("Coronavirus")
  emergency_dates("Codronavilus")
}
}
}
\seealso{
\code{\link[lubridate]{ymd}}
}
