% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_indicator_schema}
\alias{generate_indicator_schema}
\title{Generate commmon schema for indicator}
\usage{
generate_indicator_schema(
  .data,
  indicator_id,
  indicator_name,
  aggregation_type,
  outbreak_starting_date,
  ...
)
}
\description{
Generate commmon schema for indicator
}
\keyword{internal}
