% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_country_id_from_name}
\alias{get_country_id_from_name}
\title{mapping from country name to id (this is used for frontend purposes)}
\usage{
get_country_id_from_name(country_name)
}
\description{
mapping from country name to id (this is used for frontend purposes)
}
\keyword{internal}
