% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_emergency_id_from_date}
\alias{get_emergency_id_from_date}
\title{Gets Emergency id from outbreak date}
\usage{
get_emergency_id_from_date(outbreak_starting_date)
}
\description{
Gets Emergency id from outbreak date
}
\keyword{internal}
