% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_emergency_name_from_date}
\alias{get_emergency_name_from_date}
\title{Desume Emergency name from its date}
\usage{
get_emergency_name_from_date(outbreak_starting_date)
}
\description{
Desume Emergency name from its date
}
\keyword{internal}
