% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10-contracts-with-mods.R
\name{ind_10}
\alias{ind_10}
\title{Compute Contracts with modifications}
\usage{
ind_10(data, publication_date, stat_unit, id_variants, emergency_name)
}
\arguments{
\item{data}{mock bndcp data}

\item{publication_date}{PARAM_DESCRIPTION}

\item{stat_unit}{statistical unit of measurement, aggregation variable, the indicator target}

\item{id_variants}{the variants key id value}

\item{emergency_name}{emergency name character string for which you want to evaluate the indicator, e.g. "Coronavirus" "Terremoto Aquila"}
}
\value{
indicator schema as from \code{generate_indicator_schema()} rows determined by aggregation level and \code{indicator_value} based on statistical test performed in \code{ind_10}
}
\description{
The indicator reveals the fraction of contracts with at least one modification communication.
}
\examples{
\dontrun{
if (interactive()) {
  data("mock_data_core")
  ind_10(
    data = mock_data_core,
    publication_date = data_pubblicazione,
    stat_unit = provincia,
    id_variants = id_variante,
    emergency_name = "coronavirus"
  )
}
}
}
\seealso{
\code{\link[lubridate]{ymd}}
\code{\link[dplyr]{mutate}}, \code{\link[dplyr]{if_else}}, \code{\link[dplyr]{group_by}}, \code{\link[dplyr]{summarise}}, \code{\link[dplyr]{context}}, \code{\link[dplyr]{filter}}, \code{\link[dplyr]{c("rowwise", "rowwise", "rowwise")}}, \code{\link[dplyr]{select}}, \code{\link[dplyr]{reexports}}, \code{\link[dplyr]{rename}}
\code{\link[forcats]{as_factor}}
\code{\link[stats]{prop.test}}
}
