% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-awd-eco-value-across-crisis.R
\name{ind_2}
\alias{ind_2}
\title{Compute Awarded economic value across the crisis indicator}
\usage{
ind_2(
  data,
  contract_value,
  publication_date,
  emergency_name,
  stat_unit,
  test_type
)
}
\arguments{
\item{data}{test bndcp data}

\item{contract_value}{Overall Amount of the Tender}

\item{publication_date}{the date in which the tender was published}

\item{emergency_name}{emergency name character string for which you want to evaluate the indicator, e.g. "Coronavirus" "Terremoto Aquila"}

\item{stat_unit}{statistical target unit of measurement, aggregation variable, the indicator target}

\item{test_type}{character vector identifying the type of test you want to execute, alternatives are c("ks", "wilcoxon")}
}
\value{
indicator schema as from \code{generate_indicator_schema()} rows determined by aggregation level and \code{indicator_value} based on statistical test performed in \code{ind_2}
}
\description{
The indicator focuses on companies that after the emergency outbreak were awarded public contracts much higher in economic value than before the emergency
\subsection{Motivation:}{

The red flag considers at risk those companies that \strong{exceptionally increase} their \emph{competitive power} over the outbreak, in terms of economic value of their awarded contracts on the relevant economic market.
}

\subsection{Scoring Rule}{

If Test-statistic significant-> 1, otherwise -> 0
}

\subsection{Main target unit}{

This indicator targets \strong{Companies}
}
}
\examples{
\dontrun{
if (interactive()) {
  data("mock_data_core")
  ind_2(
    data = mock_data_core,
    contract_value = importo_complessivo_gara,
    publication_date = data_pubblicazione,
    stat_unit = provincia,
    test_type = "ks",
    emergency_name = "coronavirus"
  )
}
}
}
\seealso{
\code{\link[lubridate]{ymd}}
\code{\link[dplyr]{mutate}}, \code{\link[dplyr]{if_else}}, \code{\link[dplyr]{group_by}}, \code{\link[dplyr]{filter}}, \code{\link[dplyr]{summarise}}
\code{\link[forcats]{as_factor}}
\code{\link[stringr]{str_sub}}
\code{\link[tidyr]{nest}}, \code{\link[tidyr]{drop_na}}
}
