% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-ec-dev-across-crisis.R
\name{ind_3}
\alias{ind_3}
\title{Compute Economic deviation across the crisis indicator}
\usage{
ind_3(
  data,
  award_value,
  sums_paid,
  stat_unit,
  emergency_name,
  publication_date,
  test_type,
  cpvs,
  ...
)
}
\arguments{
\item{data}{This argument should be a data frame or tibble containing the data you want to use to calculate the indicator.}

\item{award_value}{This argument corresponds to the name of the column in data containing award value of the tender.}

\item{sums_paid}{This argument corresponds to the name of the column in data containing the amount paid by the contracting authority for each contract. The values in this column should be numeric.}

\item{stat_unit}{This argument should be a character string specifying the statistical unit of measurement or aggregation variable for the indicator. In this indicator both companies and contracting authorities are the targets.}

\item{emergency_name}{This argument should be a character string specifying the name of the emergency or event you are analyzing. Examples could include "Coronavirus" or "Terremoto Aquila".}

\item{publication_date}{This argument corresponds to the name of the column in data containing the publication date for each notice or report.}

\item{test_type}{This argument should be a character vector specifying the type of hypothesis test (belonging to category 2 i.e. see statistical_tests.R) to apply to the data. Available options are "wilcoxon" e "ks".}

\item{cpvs}{character vector of macro-cpv on which data is filtered out. A panel of experts have already chosen which cpvs are most affected by which emergency for you.}

\item{...}{other parameters to pass to \code{generate_indicator_schema} as \code{country_name} if that not Italy, which is default behavior.}
}
\value{
indicator schema as from \code{generate_indicator_schema}
}
\description{
The indicator reveals whether there has been a deviation of the contract actual execution economic value from its initial awarded value.
\subsection{Motivation:}{

The red flag considers at risk companies whose contracts undergo a \strong{significant increase} of their \emph{economic deviation ratio} - i.e., ratio between awarded economic value and actual amount paid - across the crisis.
}

\subsection{Scoring Rule}{

The output will give $1 - pvalue$, which will then be dichotomised to 1 if statistical test is significant, 0 otherwise.
}

\subsection{Main target unit}{

This indicator targets \strong{companies} and \strong{contracting authorities}
}
}
\examples{
\dontrun{
if (interactive()) {
  ind_3(
    data = mock_data_core,
    publication_date = data_pubblicazione,
    award_value = importo_aggiudicazione,
    sums_paid = importo_lotto,
    stat_unit = cf_amministrazione_appaltante,
    emergency_name = "coronavirus"
  )
}
}
}
\seealso{
\code{\link[lubridate]{ymd}}
\code{\link[dplyr]{filter}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{if_else}}, \code{\link[dplyr]{group_by}}, \code{\link[dplyr]{summarise}}, \code{\link[dplyr]{context}}
}
