% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-len-dev-across-crisis.R
\name{ind_4}
\alias{ind_4}
\title{Compute Contract length deviation across the crisis indicator}
\usage{
ind_4(
  data,
  exp_end,
  eff_start,
  eff_end,
  stat_unit,
  emergency_name,
  publication_date,
  test_type,
  cpvs,
  ...
)
}
\arguments{
\item{data}{a dataframe containing the data to use for computing the indicator.}

\item{exp_end}{name of the variable in \code{data} containing the expected end date of each contract, i.e., contract completion date.}

\item{eff_start}{name of the variable in \code{data} containing the effective start date of each contract.}

\item{eff_end}{name of the variable in \code{data} containing the contract execution effective end date.}

\item{stat_unit}{name of the variable in \code{data} containing the target unit ID (in this case, company or contracting authority).}

\item{emergency_name}{string specifying the name of the emergency to consider. Examples could include "Coronavirus" or
"Terremoto Centro Italia 2016-2017".}

\item{publication_date}{name of the variable in \code{data} containing the publication date of each contract.}

\item{test_type}{string specifying the statistical test to use for computing the indicator. Available options are "wilcoxon"
and "ks" (Kolmogorov-Smirnov test).}

\item{cpvs}{character vector of CPV divisions (first two digits of CPV code) on which \code{data} are filtered out.
Note: a panel of experts have already chosen which CPV divisions are most affected by which emergency.}

\item{...}{other parameters to pass to \code{\link[=generate_indicator_schema]{generate_indicator_schema()}}, such as \code{country_name} (default: Italy).}
}
\value{
indicator schema as from \code{\link[=generate_indicator_schema]{generate_indicator_schema()}}.
}
\description{
The indicator reveals whether there has been an increase in the deviation of the contract actual execution duration from its
stated/expected duration.
\subsection{Motivation}{

The red flag considers at risk companies/contracting authorities whose contracts undergo a significant increase of their
\emph{length deviation ratio} - i.e., ratio between contract actual execution duration and expected duration - across the crisis.
}

\subsection{Scoring rule}{

The computation procedure returns \emph{1 - p-value} of the involved test (so that high values of the indicator correspond
to high levels of corruption risk). When computing the composite, it will be dichotomised to 1 if statistical test is
significant, and 0 otherwise (see \code{\link[=normalise]{normalise()}}).
}

\subsection{Main target unit}{

This indicator targets \strong{companies} and \strong{contracting authorities}.
}
}
\examples{
\dontrun{
if (interactive()) {
  ind_4(
    data = mock_data_core,
    publication_date = data_pubblicazione,
    exp_end = data_termine_contrattuale,
    eff_end = data_effettiva_ultimazione,
    eff_start = data_inizio_effettiva,
    stat_unit = cf_amministrazione_appaltante,
    emergency_name = "coronavirus"
  )
}
}
}
\seealso{
\code{\link[lubridate]{ymd}}
\code{\link[dplyr]{filter}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{if_else}}, \code{\link[dplyr]{group_by}}, \code{\link[dplyr]{summarise}}, \code{\link[dplyr]{context}}
}
