% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-win-share-issuer-across-crisis.R
\name{ind_5}
\alias{ind_5}
\title{Compute Winner's share of issuer's contract across the crisis indicator}
\usage{
ind_5(data, stat_unit, publication_date, winners, emergency_name)
}
\arguments{
\item{data}{data to be passed, expects tibble}

\item{stat_unit}{Column of entities who decide who win the public procurement}

\item{publication_date}{The date when the tender was published}

\item{winners}{Column of winning companies}

\item{emergency_name}{emergency name character string for which you want to evaluate the indicator, e.g. "Coronavirus" "Terremoto Aquila"}
}
\value{
indicator schema as from \code{generate_indicator_schema}
}
\description{
Compute ratio of Gini of winning companies for agency, before and after a crisis
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  ind_5(
    data = mock_data_core,
    publication_date = data_pubblicazione,
    stat_unit = cf_amministrazione_appaltante,
    winners = denominazione,
    emergency_name = "coronavirus"
  )
}
}
}
\seealso{
\code{\link[lubridate]{ymd}}
\code{\link[dplyr]{filter}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{if_else}}, \code{\link[dplyr]{group_by}}, \code{\link[dplyr]{summarise}}, \code{\link[dplyr]{context}}
\code{\link[forcats]{as_factor}}
}
