% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-os-opportunistic-company.R
\name{ind_7}
\alias{ind_7}
\title{Compute One-shot opportunistic companies over the crisis indicator}
\usage{
ind_7(data, final_award_date, emergency_name, stat_unit, years_before, ...)
}
\arguments{
\item{data}{a dataframe containing the data to use for computing the indicator.}

\item{final_award_date}{name of the variable in \code{data} containing award notice date for each contract.}

\item{emergency_name}{string specifying the name of the emergency to consider. Examples could include "Coronavirus" or
"Terremoto Centro Italia 2016-2017".}

\item{stat_unit}{name of the variable in \code{data} containing the target unit ID (in this case, company).}

\item{years_before}{how many years before the contract date we need to look for the presence of awards to a specific company.}

\item{...}{other parameters to pass to \code{\link[=generate_indicator_schema]{generate_indicator_schema()}}, such as \code{country_name} (default: Italy).}
}
\value{
indicator schema as from \code{\link[=generate_indicator_schema]{generate_indicator_schema()}}.
}
\description{
The indicator focuses on companies that after the emergency outbreak win one or more public contracts,
without winning in the years before the emergency outbreak (e.g., in the preceding five years), with reference
to contracts belonging to relevant economic market.
\subsection{Motivation}{

The red flag considers at risk those companies that show a "one-shot opportunistic behaviour", that is,
companies that after the emergency outbreak have been awarded one or more public contracts but have not shown
any competitive power in the previous year(s).
}

\subsection{Scoring rule}{

If a company wins one or more contracts after the emergency outbreak but has not won any contracts in the years
before the emergency outbreak, the indicator will be equal to 1; otherwise, it will be equal to 0.
}

\subsection{Main target unit}{

This indicator targets \strong{companies}.
}
}
\examples{
\dontrun{
if (interactive()) {
  mock_data_core <- mock_data_core |>
    tidyr::unnest(aggiudicatari, keep_empty = TRUE)
  ind_7(
    data = mock_data_core,
    final_award_date = data_aggiudicazione_definitiva,
    stat_unit = codice_fiscale,
    emergency_name = "coronavirus",
    years_before = 1
  )
}
}
}
\seealso{
\code{\link[lubridate]{ymd}}, \code{\link[lubridate]{interval}}, \code{\link[lubridate]{period}}
\code{\link[tidyr]{nest}}
\code{\link[dplyr]{mutate}}, \code{\link[dplyr]{across}}, \code{\link[dplyr]{reexports}}, \code{\link[dplyr]{if_else}}, \code{\link[dplyr]{group_by}}
}
