% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-pre-exist-contracts-ext-after-emerg-outb.R
\name{ind_8}
\alias{ind_8}
\title{Compute Pre-existing contracts modified after the crisis indicator}
\usage{
ind_8(
  data,
  publication_date,
  stat_unit,
  variant_date,
  emergency_name,
  months_win = 6,
  ...
)
}
\arguments{
\item{data}{a dataframe containing the data to use for computing the indicator.}

\item{publication_date}{name of the variable in \code{data} containing the publication date of each contract.}

\item{stat_unit}{name of the variable in \code{data} containing the target unit ID (in this case, company or contracting authority).}

\item{variant_date}{name of the variable in \code{data} containing the date of occurrence for each contract variant.}

\item{emergency_name}{string specifying the name of the emergency to consider. Examples could include "Coronavirus" or
"Terremoto Centro Italia 2016-2017".}

\item{months_win}{parametric time window (in months) from the emergency outbreak. Only contract modifications made after
this window are considered.}

\item{...}{other parameters to pass to \code{\link[=generate_indicator_schema]{generate_indicator_schema()}}, such as \code{country_name} (default: Italy).}
}
\value{
indicator schema as from \code{\link[=generate_indicator_schema]{generate_indicator_schema()}}.
}
\description{
The indicator reveals whether a contracting authority/company has at least one pre-existing contract with respect
to the emergency outbreak (i.e., with a call before the beginning of the emergency), which has been modified after
the emergency outbreak, with reference to relevant economic market.
\subsection{Motivation}{

The red flag considers at risk the contracting authority/company with existing contracts before the emergency outbreak
and subsequently modified through variants after the emergency began; this assessment excludes modifications made immediately
thereafter, using a parametric window (e.g., within the following six months).
}

\subsection{Scoring rule}{

For a given target unit, if a contract existing before the outbreak is modified through variants after the emergency outbreak
(given the parametric window, e.g., six months), the indicator will be equal to 1; otherwise, it will be equal to 0.
}

\subsection{Main target unit}{

This indicator targets \strong{companies} and \strong{contracting authorities}.
}
}
\examples{
\dontrun{
if (interactive()) {
  mock_data_core <- mock_data_core |>
    tidyr::unnest(varianti, keep_empty = TRUE)
  ind_8(
    data = mock_data_core,
    publication_date = data_pubblicazione,
    stat_unit = cf_amministrazione_appaltante,
    variant_date = data_approvazione_variante,
    months_win = 6,
    emergency_name = "coronavirus"
  )
}
}
}
\seealso{
\code{\link[dplyr]{mutate}}, \code{\link[dplyr]{if_else}}, \code{\link[dplyr]{group_by}}, \code{\link[dplyr]{summarise}}, \code{\link[dplyr]{context}}, \code{\link[dplyr]{n_distinct}}, \code{\link[dplyr]{rename}}
\code{\link[lubridate]{ymd}}, \code{\link[lubridate]{interval}}
}
