% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-lengthy-contract.R
\name{ind_9}
\alias{ind_9}
\title{Compute Lenghty contracts indicator}
\usage{
ind_9(
  data,
  publication_date,
  stat_unit,
  eff_start,
  eff_end,
  emergency_name,
  ...
)
}
\arguments{
\item{data}{a dataframe containing the data to use for computing the indicator.}

\item{publication_date}{name of the variable in \code{data} containing the publication date of each contract.}

\item{stat_unit}{name of the variable in \code{data} containing the target unit ID (in this case, company or contracting authority).}

\item{eff_start}{name of the variable in \code{data} containing the effective start date of each contract.}

\item{eff_end}{name of the variable in \code{data} containing the contract execution effective end date.}

\item{emergency_name}{string specifying the name of the emergency to consider. Examples could include "Coronavirus" or
"Terremoto Centro Italia 2016-2017".}

\item{...}{other parameters to pass to \code{\link[=generate_indicator_schema]{generate_indicator_schema()}}, such as \code{country_name} (default: Italy).}
}
\value{
indicator schema as from \code{\link[=generate_indicator_schema]{generate_indicator_schema()}}.
}
\description{
The indicator reveals whether the (average) contract duration of a contracting authority/company is significantly longer
than the overall (average) contract duration, with reference to post-emergency contracts belonging to relevant economic market.
\subsection{Motivation}{

The red flag considers at risk contracting authorities/companies with post-emergency contracts whose excessive duration
is not justified by the nature of the crisis, that is, contracts longer than the overall average duration of contracts won
after the emergency outbreak.
}

\subsection{Scoring rule}{

The computation procedure returns \emph{1 - p-value} of the involved test - Wilcoxon test in this case - (so that high values of
the indicator correspond to high levels of corruption risk). When computing the composite,
it will be dichotomised to 1 if statistical test is significant, and 0 otherwise (see \code{\link[=normalise]{normalise()}}).
}

\subsection{Main target unit}{

This indicator targets \strong{companies} and \strong{contracting authorities}.
}
}
\examples{
\dontrun{
if (interactive()) {
  ind_9(
    data = mock_data_core,
    publication_date = data_pubblicazione,
    stat_unit = cf_amministrazione_appaltante,
    eff_start = data_inizio_effettiva,
    eff_end = data_effettiva_ultimazione,
    emergency_name = "coronavirus"
  )
}
}
}
\seealso{
\code{\link[dplyr]{mutate}}, \code{\link[dplyr]{if_else}}, \code{\link[dplyr]{filter}}, \code{\link[dplyr]{pull}}, \code{\link[dplyr]{group_by}}, \code{\link[dplyr]{summarise}}, \code{\link[dplyr]{context}}, \code{\link[dplyr]{rename}}
\code{\link[lubridate]{ymd}}
\code{\link[forcats]{as_factor}}
\code{\link[stringr]{str_sub}}
\code{\link[rlang]{as_string}}
}
