% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composite.R
\name{ind_all_companies}
\alias{ind_all_companies}
\title{Elementary indicators for companies}
\usage{
ind_all_companies(data, data_ind8, emergency_name)
}
\description{
\code{ind_all_companies} computes all the elementary indicators with companies as target unit,
by sequentially calling \code{\link[=ind_1]{ind_1()}}, \code{\link[=ind_2]{ind_2()}}, \code{\link[=ind_3]{ind_3()}}, \code{\link[=ind_4]{ind_4()}}, \code{\link[=ind_7]{ind_7()}}, \code{\link[=ind_8]{ind_8()}} and \code{\link[=ind_9]{ind_9()}}.
It is an internal function.
}
\keyword{internal}
