% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composite.R
\name{ind_all_contr_auth}
\alias{ind_all_contr_auth}
\title{Elementary indicators for contracting authorities}
\usage{
ind_all_contr_auth(data, data_ind8, emergency_name)
}
\description{
\code{ind_all_contr_auth} computes all the elementary indicators with contracting authorities
as target unit, by sequentially calling \code{\link[=ind_3]{ind_3()}}, \code{\link[=ind_4]{ind_4()}}, \code{\link[=ind_5]{ind_5()}}, \code{\link[=ind_6]{ind_6()}}, \code{\link[=ind_8]{ind_8()}} and \code{\link[=ind_9]{ind_9()}}.
It is an internal function.
}
\keyword{internal}
