% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{italian_aggregation_mapping}
\alias{italian_aggregation_mapping}
\title{italian mapping from aggregation_id to aggregation_name}
\format{
An object of class \code{list} of length 4.
}
\usage{
italian_aggregation_mapping
}
\description{
italian mapping from aggregation_id to aggregation_name
}
\keyword{internal}
