% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composite.R
\name{manage_missing}
\alias{manage_missing}
\title{Manage missing values in elementary indicators}
\usage{
manage_missing(data, missing = 0, verbose = TRUE, seed = 1234)
}
\arguments{
\item{data}{data matrix of elementary indicators (as returned by \code{\link[=create_indicator_matrix]{create_indicator_matrix()}}).}

\item{missing}{method for imputing missing values:
\itemize{
\item \code{missing = 0}: missing values are replaced with '0' (not at risk);
\item \code{missing = 1}: missing values are imputed using logistic regression. See Details.
}}

\item{verbose}{whether a summary of imputed values has to be printed (when \code{missing = 1}).}

\item{seed}{seed for the random draw from a Bernoulli r.v. (when \code{missing = 1}). See Details.}
}
\value{
data matrix with no missing value.
}
\description{
\code{manage_missing} deals with the imputation of missing values in the elementary indicators.
\strong{It is an internal function.}
}
\details{
When \code{missing = 1}, elementary indicators are split up into two groups, according to
the presence/absence of (at least one) missing values. Hence, the set of indicators without
missing values (taken as covariates) is used to 'predict' the missing values in each of the other
indicators (seen as dependent variable), using several logistic regression models, one for each
indicator with missing values. As such, these models can predict a probability for each combination
of observed indicators; then, a random draw from a Bernoulli distribution with the predicted
probability as parameter is performed for imputing '0' or '1' in the indicator with missing values.
}
\seealso{
\code{\link[tidyr]{drop_na}}
\code{\link[dplyr]{select}}
}
