% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composite.R
\name{normalise}
\alias{normalise}
\title{normalise}
\usage{
normalise(data, method = "binary", cutoff = 0.95)
}
\arguments{
\item{data}{data matrix of elementary indicators (as returned by \code{\link[=create_indicator_matrix]{create_indicator_matrix()}})}

\item{method}{normalisation method, to be chosen among:
\itemize{
\item \code{"binary"}: each elementary indicator is dichotomised (0/1) using a suitable threshold,
to be specified through the argument \code{cutoff}. Specifically, the normalised indicator
will be equal to 1 if the original indicator is greater then the threshold, and 0 otherwise;
\item \code{"ranking"}: each elementary indicator is normalised according to the ranking (see \code{\link[=rank]{rank()}});
\item \code{"z-score"}: each elementary indicator is standardised into z-scores (see \code{\link[=scale]{scale()}}).
Let \eqn{x_{qc}} be the original value of elementary indicator \eqn{q} for target unit \eqn{c}. Then,
the z-score is obtained as follows:
}

\deqn{I_{qc} = \frac{x_{qc} - \mu_q(x_{qc})}{\sigma_q(x_{qc})}}
\itemize{
\item \code{"minmax"}: each elementary indicator is normalised using the 'min-max' criterion:
}

\deqn{I_{qc} = \frac{x_{qc} - min(x_{qc})}{(max(x_{qc}) - min(x_{qc}))}}
\itemize{
\item \code{"distref"}: each elementary indicator is normalised by dividing it by its maximum;
\item \code{"catscale"}: each elementary indicator is discretised into five categories, according to suitable
sample quantiles.
}}

\item{cutoff}{threshold for dichotomising the indicators (when \code{method = "binary"}).}
}
\value{
data matrix of normalised indicators according to the chosen method.
}
\description{
\code{normalise} normalises the elementary indicators using a suitable normalisation method
(e.g., ranking, min-max, dichotomisation, etc.).
}
\details{
Normalise the elementary indicators

In the CO.R.E. project, according to the proposed set of elementary indicators,
a suitable normalisation method should be the \emph{'dichotomisation'} (i.e., \code{method = "binary"}),
in order to make all the indicators as binary. In particular, if a given normalised elementary
indicator is equal to 1, this means that the target unit is considered \strong{at risk} on the basis
of that indicator. On the other hand, the target unit is considered \strong{not at risk}.

In fact, some elementary indicators perform a statistical test and return, as risk metric, one minus
the p-value of the test (hence, a continuous scale, even if bounded in \eqn{[0,1]}); some others
do not consider any statistical tests, are binary by nature and directly return the target unit as
'not at risk/at risk' (0/1).

In order to bring all the elementary indicators to the same metric, that is, 'not at risk/at risk' (0/1),
the group of indicators that rely on statistical testing must be normalised (i.e., dichotomised)
by specifying a suitable threshold for the significance of the involved tests.
It corresponds to one minus the threshold for the significance of the p-value of the test performed
by the indicator. Example: \code{cutoff = 0.95} means that the threshold for the significance of the p-values
is the usual 0.05.
}
\examples{
\dontrun{
if(interactive()){
 normalise(data_matrix, method = "binary", cutoff = 0.95)
 }
}
}
