% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_tests.R
\name{test_set_1}
\alias{test_set_1}
\title{switch test wrt statistical circumstances set 1 (indicators 1, ...)}
\usage{
test_set_1(a, b, c, d, test_type)
}
\description{
switch test wrt statistical circumstances
}
\keyword{internal}
