% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_tests.R
\name{test_set_2}
\alias{test_set_2}
\title{switch test wrt statistical circumstances set 2 (indicators 2,3,4 ...)}
\usage{
test_set_2(data, var, group, test_type)
}
\description{
switch test wrt statistical circumstances
}
\keyword{internal}
