% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-winning-rate-across-crisis.R
\name{compute_barnard}
\alias{compute_barnard}
\title{compute Barnard_2 test in dplyr pipeline in other fashion}
\usage{
compute_barnard(a, b, c, d)
}
\description{
compute Barnard test pvalue and estimate in piped expression
}
\keyword{internal}
