% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-high-winning-rate.R
\name{compute_fisher}
\alias{compute_fisher}
\title{compute fisher test in dplyr pipeline}
\usage{
compute_fisher(a, b, c, d)
}
\description{
compute fisher test pvalue and estimate in piped expression
}
\keyword{internal}
