% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-awd-eco-value-across-crisis.R
\name{compute_wilcox}
\alias{compute_wilcox}
\title{compute Wilcoxon-Mann-Whitney test in dplyr https://it.wikipedia.org/wiki/Test_di_Wilcoxon-Mann-Whitney}
\usage{
compute_wilcox(data, var, group, exact = TRUE, alternative = "greater")
}
\description{
compute Wilcoxon-Mann-Whitney test pvalue
}
\keyword{internal}
