% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coresoi-package.R
\docType{package}
\name{coresoi-package}
\alias{coresoi}
\alias{coresoi-package}
\title{coresoi: What the Package Does (One Line, Title Case)}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://core-forge.github.io/coresoi/}
}

}
\author{
\strong{Maintainer}: Niccolò Salvini \email{niccolo.salvini27@gmail.com} (\href{https://orcid.org/0000-0002-3444-4094}{ORCID})

}
\keyword{internal}
