% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composite.R
\name{create_indicator_matrix}
\alias{create_indicator_matrix}
\title{Create matrix of elementary indicators}
\usage{
create_indicator_matrix(out_list)
}
\arguments{
\item{out_list}{list of outputs about each indicator computable for the target unit
(e.g., company or contracting authority), as returned by \code{\link[=ind_all]{ind_all()}}.}
}
\value{
data matrix with aggregation ID of the target units as first column and
indicator values as subsequent columns (according to \code{out_list}).
}
\description{
\code{create_indicator_matrix} creates the data matrix of elementary indicators
(row = target unit; columns = indicator values).
}
\details{
Target unit ID in each output of \code{out_list} can be different and a full join is carried out for merging all the
indicators and building the final data matrix.
}
\examples{
\dontrun{
if (interactive()) {
  # sample of 100k contracts
  set.seed(12345)
  i <- sample(1:nrow(mock_data_core), size = 1e5)
  mock_sample0 <- mock_data_core[sort(i), ]

  # indicators for companies
  mock_sample <- tidyr::unnest(mock_sample0, aggiudicatari, keep_empty = TRUE)
  mock_sample_variants <- tidyr::unnest(mock_sample, varianti, keep_empty = TRUE)

  out_companies <- ind_all(
    data = mock_sample,
    data_ind8 = mock_sample_variants,
    emergency_name = "coronavirus",
    target_unit = "companies"
  )
  indicator_data_matrix <- create_indicator_matrix(out_companies)
}
}
}
\keyword{internal}
