% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composite.R
\name{create_indicator_matrix}
\alias{create_indicator_matrix}
\title{create_indicator_matrix}
\usage{
create_indicator_matrix(out_list)
}
\arguments{
\item{out_list}{list of outputs about each indicator computable for the target unit
(e.g., company or contracting authority), as returned by \code{\link[=ind_all]{ind_all()}}.}
}
\value{
data matrix with aggregation ID of the target units as first column and
indicator values as subsequent columns (according to \code{out_list}).
}
\description{
\code{create_indicator_matrix} creates the data matrix of elementary indicators
(row = target unit; columns = indicator values). \strong{It is an internal function.}
}
\details{
Create matrix of elementary indicators

Target unit ID in each output of \code{out_list} can be different and a full join is carried out for merging all the
indicators and building the final data matrix.
}
\keyword{internal}
