% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_indicator_schema}
\alias{generate_indicator_schema}
\title{generate indicator schema}
\usage{
generate_indicator_schema(
  .data,
  indicator_id,
  emergency,
  indicator_name,
  missing_cols = c("codice_regione", "provincia_codice", "citta_codice",
    "cf_amministrazione_appaltante", "codice_fiscale"),
  ...
)
}
\description{
generate indicator schema
}
\keyword{internal}
