% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_indicator_schema}
\alias{generate_indicator_schema}
\title{generate indicator schema}
\usage{
generate_indicator_schema(
  .data,
  indicator_id,
  aggregation_type,
  emergency,
  indicator_name,
  country_name = "Italy",
  ...
)
}
\description{
generate indicator schema
}
\keyword{internal}
