% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_associated_cpv_from_emergency}
\alias{get_associated_cpv_from_emergency}
\title{Get the associated CPV code from emergency name}
\usage{
get_associated_cpv_from_emergency(emergency_name)
}
\arguments{
\item{emergency_name}{PARAM_DESCRIPTION}
}
\value{
cpv_code
}
\description{
Not all the emergency are going to impact each contract. Some of them like coronavirus are going to impact more CPV 33 "Medical equipments, pharmaceuticals and personal care products" insted of say "Agricultural, forestry, horticultural, aquacultural and apicultural services". Here we emplyed a set of domain experts which have worked on associating each emergency to a set of respective CPVs.
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
