% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-high-winning-rate.R
\name{ind_1}
\alias{ind_1}
\title{Compute indicator High Winning Rate}
\usage{
ind_1(
  data,
  publication_date,
  cpv,
  outbreak_starting_date = lubridate::ymd("2017-06-30"),
  divison,
  group
)
}
\arguments{
\item{data}{bndcp data}

\item{publication_date}{The date when the tender was published}

\item{cpv}{Common Procurement Vocabulary. The main vocabulary is based on a tree structure made up with codes of up to 9 digits (an 8 digit code plus a check digit). This combination of digits is associated with a wording that describes the type of supplies, works or services defining the subject of the contract}

\item{outbreak_starting_date}{the date of the emergency outbreak, Default: lubridate::ymd("2017-06-30")}

\item{divison}{first two digits in cpv code identifying the division, for more info check https://simap.ted.europa.eu/it/cpv}

\item{group}{statistical unit}
}
\value{
tibble df with 5 columns (group, prop_test, correct_prop_test, fisher_test, fisher_estimate )
}
\description{
it computes the indicator n° 1
}
\examples{
\dontrun{
if (interactive()) {
  data("test_data_bndcp_core")
  ind_1(data = test_data_bndcp_core, publication_date = data_pubblicazione, cpv = cod_cpv, group = cod_provincia2)
}
}
}
\seealso{
\code{\link[lubridate]{ymd}}
\code{\link[dplyr]{mutate}}, \code{\link[dplyr]{if_else}}, \code{\link[dplyr]{group_by}}, \code{\link[dplyr]{summarise}}, \code{\link[dplyr]{context}}, \code{\link[dplyr]{filter}}, \code{\link[dplyr]{c("rowwise", "rowwise", "rowwise")}}, \code{\link[dplyr]{select}}, \code{\link[dplyr]{reexports}}
\code{\link[stringr]{str_sub}}
\code{\link[tidyr]{nest}}
}
