% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-winning-rate-across-crisis.R
\name{ind_1}
\alias{ind_1}
\title{Compute Winning rate across the crisis indicator}
\usage{
ind_1(data, publication_date, emergency_name, stat_unit, test_type)
}
\arguments{
\item{data}{bndcp data}

\item{publication_date}{date of publication of the notice}

\item{emergency_name}{emergency name character string for which you wish to calculate the indicator for, e.g. "Coronavirus" "Terremoto Aquila"}

\item{stat_unit}{statistical unit of measurement, aggregation variable, the indicator target}

\item{test_type}{character vector string to identifying the test type you want to apply, available alternatives are c("barnard", "fisher", "z-test")}
}
\value{
indicator schema as from \code{\link[=generate_indicator_schema]{generate_indicator_schema()}}
}
\description{
The indicator focuses on companies that after the emergency outbreak were awarded public contracts much more frequently than before the emergency.
\subsection{Motivation:}{

The red flag considers at risk companies that \strong{exceptionally increase} their competitive power over the emergency outbreak, in terms of proportion of awarded contracts on the relevant economic market.
}

\subsection{Scoring Rule}{

If Test-statistic significant-> 1, otherwise -> 0
}

\subsection{Main target unit}{

This indicator targets \strong{Companies}
}
}
\examples{
\dontrun{
if (interactive()) {
  ind_1(
    data = mock_data_core,
    publication_date = data_pubblicazione,
    stat_unit = cf_amministrazione_appaltante,
    emergency_name = "coronavirus",
    test_type = "fisher"
  )
}
}
}
\seealso{
\code{\link[lubridate]{ymd}}
\code{\link[dplyr]{mutate}}, \code{\link[dplyr]{if_else}}, \code{\link[dplyr]{group_by}}, \code{\link[dplyr]{summarise}}, \code{\link[dplyr]{context}}, \code{\link[dplyr]{filter}}, \code{\link[dplyr]{"rowwise"}}, \code{\link[dplyr]{select}}, \code{\link[dplyr]{reexports}}
\code{\link[stringr]{str_sub}}
\code{\link[tidyr]{nest}}
}
