% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-winning-rate-across-crisis.R
\name{ind_1}
\alias{ind_1}
\title{Compute Winning rate across the crisis indicator}
\usage{
ind_1(data, publication_date, emergency_name, stat_unit, test_type, cpvs, ...)
}
\arguments{
\item{data}{This argument should be a data frame or tibble containing the data you want to use to calculate the indicator.}

\item{publication_date}{This argument corresponds to the name of the column in data containing the publication date for each notice or report.}

\item{emergency_name}{This argument should be a character string specifying the name of the emergency or event you are analyzing. Examples could include "Coronavirus" or "Terremoto Aquila".}

\item{stat_unit}{This argument should be a character string specifying the statistical unit of measurement or aggregation variable for the indicator. In this indicator companies are the targets}

\item{test_type}{This argument should be a character vector specifying the type of hypothesis test (belonging to category 1 i.e. see statistical_tests.R) to apply to the data. Available options are "barnard", "fisher", or "z-test".}

\item{cpvs}{character vector of macro-cpv on which data is filtered out. A panel of experts have already chosen which cpvs are most affected by which emergency for you.}

\item{...}{other parameters to pass to \code{generate_indicator_schema} as \code{country_name} if that not Italy, which is default behavior.}
}
\value{
indicator schema as from \code{\link[=generate_indicator_schema]{generate_indicator_schema()}}
}
\description{
The indicator focuses on companies that after the emergency outbreak were awarded public contracts much more frequently than before the emergency.
\subsection{Motivation:}{

The red flag considers at risk companies that \strong{exceptionally increase} their competitive power over the emergency outbreak, in terms of proportion of awarded contracts on the relevant economic market.
}

\subsection{Scoring Rule:}{

The output will give $1 - pvalue$, which will then be dichotomised to 1 if statistical test is significant, 0 otherwise.
}

\subsection{Main target unit:}{

This indicator targets \strong{Companies}
}
}
\examples{
\dontrun{
if (interactive()) {
  ind_1(
    data = mock_data_core,
    publication_date = data_pubblicazione,
    stat_unit = cf_amministrazione_appaltante,
    emergency_name = "coronavirus",
    test_type = "fisher"
  )
}
}
}
\seealso{
\code{\link[lubridate]{ymd}}
\code{\link[dplyr]{mutate}}, \code{\link[dplyr]{if_else}}, \code{\link[dplyr]{group_by}}, \code{\link[dplyr]{summarise}}, \code{\link[dplyr]{context}}, \code{\link[dplyr]{filter}}, \code{\link[dplyr]{rowwise}}, \code{\link[dplyr]{select}}, \code{\link[dplyr]{reexports}}
\code{\link[stringr]{str_sub}}
\code{\link[tidyr]{nest}}
}
