% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-high-winning-rate.R
\name{ind_1}
\alias{ind_1}
\title{Compute indicator High Winning Rate}
\usage{
ind_1(data, publication_date, cpv, emergency_name, stat_unit)
}
\arguments{
\item{data}{bndcp data}

\item{publication_date}{The date when the tender was published}

\item{cpv}{Common Procurement Vocabulary. The main vocabulary is based on a tree structure made up with codes of up to 9 digits (an 8 digit code plus a check digit). This combination of digits is associated with a wording that describes the type of supplies, works or services defining the subject of the contract}

\item{emergency_name}{emergency name character string for which you want to evaluate the indicator, e.g. "Coronavirus" "Terremoto Aquila"}

\item{stat_unit}{statistical unit of measurement, aggregation variable, the indicator target}

\item{divison}{first two digits in cpv code identifying the division, for more info check https://simap.ted.europa.eu/it/cpv}
}
\value{
indicator schema as from \code{generate_indicator_schema}
}
\description{
it computes the indicator n° 1
}
\examples{
\dontrun{
if (interactive()) {
  data("test_data_bndcp_core")
  ind_1(
    data = mock_data_core,
    publication_date = data_pubblicazione,
    cpv = cod_cpv,
    stat_unit = provincia,
    emergency_name = "coronavirus"
  )
}
}
}
\seealso{
\code{\link[lubridate]{ymd}}
\code{\link[dplyr]{mutate}}, \code{\link[dplyr]{if_else}}, \code{\link[dplyr]{group_by}}, \code{\link[dplyr]{summarise}}, \code{\link[dplyr]{context}}, \code{\link[dplyr]{filter}}, \code{\link[dplyr]{c("rowwise", "rowwise", "rowwise")}}, \code{\link[dplyr]{select}}, \code{\link[dplyr]{reexports}}
\code{\link[stringr]{str_sub}}
\code{\link[tidyr]{nest}}
}
