% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11-dist-between-awd-val-sums-paid.R
\name{ind_11}
\alias{ind_11}
\title{ind_11}
\usage{
ind_11(
  data,
  award_value,
  sums_paid,
  group,
  outbreak_starting_date = lubridate::ymd("2017-06-30"),
  publication_date
)
}
\arguments{
\item{data}{dataset to be passed, expects tibble}

\item{award_value}{The date when the tender was awarded}

\item{sums_paid}{The amount paid by the C.A.}

\item{group}{the statistical unit of measurement (can be a vector of grouping variables), i.e. variable to group by}

\item{outbreak_starting_date}{The date when the emergency officially started, Default: lubridate::ymd("2017-06-30")}

\item{publication_date}{The date when the tender was published}
}
\value{
a tibble \link{n x 5} containing cf_amministrazione_appaltante
}
\description{
The difference between the sum foreseen in the contract and the actual payment by the C.A.
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  ind_11(
    .data = bncp_data, publication_date = data_pubblicazione,
    award_value = importo_aggiudicazione, sums_paid = imp_finale,
    cf_amministrazione_appaltante
  )
}
}
}
\seealso{
\code{\link[lubridate]{ymd}}
\code{\link[dplyr]{filter}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{if_else}}, \code{\link[dplyr]{group_by}}, \code{\link[dplyr]{summarise}}, \code{\link[dplyr]{context}}
\code{\link[forcats]{as_factor}}
}
