% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-awd-eco-value-across-crisis.R
\name{ind_2}
\alias{ind_2}
\title{Compute Awarded economic value across the crisis indicator}
\usage{
ind_2(
  data,
  contract_value,
  publication_date,
  emergency_name,
  stat_unit,
  test_type
)
}
\arguments{
\item{data}{This argument should be a data frame or tibble containing the data you want to use to calculate the indicator.}

\item{contract_value}{This argument corresponds to the name of the column in data containing the overall amount of the tender for each contract. The values in this column should be numeric.}

\item{publication_date}{This argument corresponds to the name of the column in data containing the publication date for each notice or report.}

\item{emergency_name}{This argument should be a character string specifying the name of the emergency or event you are analyzing. Examples could include "Coronavirus" or "Terremoto Aquila".}

\item{stat_unit}{This argument should be a character string specifying the statistical unit of measurement or aggregation variable for the indicator. In this indicator companies are the target.}

\item{test_type}{This argument should be a character vector specifying the type of hypothesis test to apply to the data. Available options are "wilcoxon" e "ks".}
}
\value{
indicator schema as from \code{generate_indicator_schema()} rows determined by aggregation level and \code{indicator_value} based on statistical test performed in \code{ind_2}
}
\description{
The indicator focuses on companies that after the emergency outbreak were awarded public contracts much higher in economic value than before the emergency
\subsection{Motivation:}{

The red flag considers at risk those companies that \strong{exceptionally increase} their \emph{competitive power} over the outbreak, in terms of economic value of their awarded contracts on the relevant economic market.
}

\subsection{Scoring Rule}{

The output will give $1 - pvalue$, which will then be dichotomised to 1 if statistical test is significant, 0 otherwise.
}

\subsection{Main target unit}{

This indicator targets \strong{Companies}
}
}
\examples{
\dontrun{
if (interactive()) {
  data("mock_data_core")
  ind_2(
    data = mock_data_core,
    contract_value = importo_complessivo_gara,
    publication_date = data_pubblicazione,
    stat_unit = provincia,
    test_type = "ks",
    emergency_name = "coronavirus"
  )
}
}
}
\seealso{
\code{\link[lubridate]{ymd}}
\code{\link[dplyr]{mutate}}, \code{\link[dplyr]{if_else}}, \code{\link[dplyr]{group_by}}, \code{\link[dplyr]{filter}}, \code{\link[dplyr]{summarise}}
\code{\link[stringr]{str_sub}}
\code{\link[tidyr]{nest}}, \code{\link[tidyr]{drop_na}}
}
