% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-awd-eco-value-across-crisis.R
\name{ind_2}
\alias{ind_2}
\title{Compute Awarded economic value across the crisis indicator}
\usage{
ind_2(data, cpv, contract_value, publication_date, emergency_name, stat_unit)
}
\arguments{
\item{data}{test bndcp data}

\item{cpv}{Common Procurement Vocabulary. The main vocabulary is based on a tree structure made up with codes of up to 9 digits (an 8 digit code plus a check digit). This combination of digits is associated with a wording that describes the type of supplies, works or services defining the subject of the contract}

\item{contract_value}{the value of the contract}

\item{emergency_name}{emergency name character string for which you want to evaluate the indicator, e.g. "Coronavirus" "Terremoto Aquila"}

\item{stat_unit}{statistical unit of measurement, aggregation variable, the indicator target}
}
\value{
indicator schema as from \code{generate_indicator_schema()} rows determined by aggregation level and \code{indicator_value} based on statistical test performed in \code{ind_2}
}
\description{
companies that after the emergency outbreak were awarded public contracts much higher in economic value than before the emergency scenario
}
\examples{
\dontrun{
if (interactive()) {
  data("mock_data_core")
  ind_2(
    data = mock_data_core,
    cpv = cod_cpv,
    contract_value = importo_complessivo_gara,
    publication_date = data_pubblicazione,
    stat_unit = provincia,
    emergency_name = "coronavirus"
  )
}
}
}
\seealso{
\code{\link[lubridate]{ymd}}
\code{\link[dplyr]{mutate}}, \code{\link[dplyr]{if_else}}, \code{\link[dplyr]{group_by}}, \code{\link[dplyr]{filter}}, \code{\link[dplyr]{summarise}}
\code{\link[forcats]{as_factor}}
\code{\link[stringr]{str_sub}}
\code{\link[tidyr]{nest}}, \code{\link[tidyr]{drop_na}}
}
