% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-high-economic-value.R
\name{ind_2}
\alias{ind_2}
\title{Compute High Economic Value indicator}
\usage{
ind_2(
  data,
  cpv,
  contract_value,
  outbreak_starting_date = lubridate::ymd("2017-06-30")
)
}
\arguments{
\item{data}{bndcp data}

\item{cpv}{Common Procurement Vocabulary. The main vocabulary is based on a tree structure made up with codes of up to 9 digits (an 8 digit code plus a check digit). This combination of digits is associated with a wording that describes the type of supplies, works or services defining the subject of the contract}

\item{contract_value}{the value of the contract}

\item{outbreak_starting_date}{the date of the emergency outbreak, Default: lubridate::ymd("2017-06-30")}
}
\value{
tibble n x 5
}
\description{
High Economic Value for statistical unit
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
\seealso{
\code{\link[lubridate]{ymd}}
\code{\link[dplyr]{mutate}}, \code{\link[dplyr]{if_else}}, \code{\link[dplyr]{group_by}}, \code{\link[dplyr]{filter}}, \code{\link[dplyr]{summarise}}
\code{\link[forcats]{as_factor}}
\code{\link[stringr]{str_sub}}
\code{\link[tidyr]{nest}}, \code{\link[tidyr]{drop_na}}
}
