% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-ec-dev-across-crisis.R
\name{ind_3}
\alias{ind_3}
\title{Compute Contract economic deviation across the crisis indicator}
\usage{
ind_3(
  data,
  award_value,
  sums_paid,
  stat_unit,
  emergency_name,
  publication_date,
  test_type,
  cpvs,
  ...
)
}
\arguments{
\item{data}{a dataframe containing the data to use for computing the indicator.}

\item{award_value}{name of the variable in \code{data} containing the award value of each contract.}

\item{sums_paid}{name of the variable in \code{data} containing the contract actual execution economic value (i.e., amount paid by the contracting authority).}

\item{stat_unit}{name of the variable in \code{data} containing the target unit ID (in this case, company or contracting authority).}

\item{emergency_name}{string specifying the name of the emergency to consider. Examples could include "Coronavirus" or "Terremoto Centro Italia 2016-2017".}

\item{publication_date}{name of the variable in \code{data} containing the publication date of each contract.}

\item{test_type}{string specifying the statistical test to use for computing the indicator. Available options are "wilcoxon" and "ks" (Kolmogorov-Smirnov test).}

\item{cpvs}{character vector of CPV divisions (first two digits of CPV code) on which \code{data} are filtered out. Note: a panel of experts have already chosen which CPV
divisions are most affected by which emergency.}

\item{...}{other parameters to pass to \code{\link[=generate_indicator_schema]{generate_indicator_schema()}}, such as \code{country_name} (default: Italy).}
}
\value{
indicator schema as from \code{\link[=generate_indicator_schema]{generate_indicator_schema()}}.
}
\description{
The indicator reveals whether there has been an increase in the deviation of the contract actual execution economic value from its initial awarded value,
with reference to contracts belonging to relevant economic market.
\subsection{Motivation}{

The red flag considers at risk companies/contracting authorities whose contracts undergo a significant increase of their \emph{economic deviation ratio} - i.e.,
ratio between actual amount paid and awarded economic value - across the crisis.
}

\subsection{Scoring rule}{

The computation procedure returns \emph{1 - p-value} of the involved test (so that high values of the indicator correspond to high levels of corruption risk). When computing the composite,
it will be dichotomised to 1 if statistical test is significant, and 0 otherwise (see \code{\link[=normalise]{normalise()}}).
}

\subsection{Main target unit}{

This indicator targets \strong{companies} and \strong{contracting authorities}.
}
}
\examples{
\dontrun{
if (interactive()) {
  ind_3(
    data = mock_data_core,
    publication_date = data_pubblicazione,
    award_value = importo_lotto,
    sums_paid = importo_finale,
    stat_unit = cf_amministrazione_appaltante,
    emergency_name = "coronavirus",
    test_type = "wilcoxon"
  )
}
}
}
\seealso{
\code{\link[lubridate]{ymd}}
\code{\link[dplyr]{filter}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{if_else}}, \code{\link[dplyr]{group_by}}, \code{\link[dplyr]{summarise}}, \code{\link[dplyr]{context}}
}
