% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-os-opportunistic-company.R
\name{ind_3}
\alias{ind_3}
\title{compute One-shot opportunistic companies}
\usage{
ind_3(
  data,
  publication_date,
  awd_company,
  outbreak_starting_date = lubridate::ymd("2017-06-30")
)
}
\arguments{
\item{data}{test bndcp data}

\item{publication_date}{The date when the tender was published}

\item{awd_company}{The unique ID Code that identifies the awarded company (ex. VAT or Tax Number)}

\item{outbreak_starting_date}{the date of the emergency outbreak, Default: lubridate::ymd("2017-06-30")}
}
\value{
tibble n x 7
}
\description{
The indicator reveals whether the company that won the public tender or contract has a previous history of at least one year since the opening of the tender.
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
\seealso{
\code{\link[lubridate]{ymd}}, \code{\link[lubridate]{character(0)}}, \code{\link[lubridate]{interval}}, \code{\link[lubridate]{period}}
\code{\link[tidyr]{nest}}
\code{\link[dplyr]{mutate}}, \code{\link[dplyr]{across}}, \code{\link[dplyr]{reexports}}, \code{\link[dplyr]{if_else}}, \code{\link[dplyr]{group_by}}
\code{\link[forcats]{as_factor}}
}
