% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-ec-dev-across-crisis.R
\name{ind_3}
\alias{ind_3}
\title{Economic deviation across the crisis indicator}
\usage{
ind_3(
  data,
  award_value,
  sums_paid,
  stat_unit,
  emergency_name,
  publication_date
)
}
\arguments{
\item{data}{data to be passed, expects tibble}

\item{award_value}{The date when the tender was awarded}

\item{sums_paid}{The amount paid by the C.A.}

\item{stat_unit}{the statistical unit of measurement (can be a vector of grouping variables), i.e. variable to group by}

\item{emergency_name}{emergency name character string for which you want to evaluate the indicator, e.g. "Coronavirus" "Terremoto Aquila"}

\item{publication_date}{The date when the tender was published}
}
\value{
indicator schema as from \code{generate_indicator_schema}
}
\description{
The difference between the sum foreseen in the contract and the actual payment by the C.A. (ita SA stazione Appaltante)
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  ind_3(
    data = mock_data_core, publication_date = data_pubblicazione,
    award_value = importo_aggiudicazione, sums_paid = importo_lotto,
    stat_unit = cf_amministrazione_appaltante,
    emergency_name = "coronavirus"
  )
}
}
}
\seealso{
\code{\link[lubridate]{ymd}}
\code{\link[dplyr]{filter}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{if_else}}, \code{\link[dplyr]{group_by}}, \code{\link[dplyr]{summarise}}, \code{\link[dplyr]{context}}
\code{\link[forcats]{as_factor}}
}
