% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-len-dev-across-crisis.R
\name{ind_4}
\alias{ind_4}
\title{Compute Contract Lenght deviation across the crisis indicator}
\usage{
ind_4(
  data,
  exp_end,
  eff_start,
  eff_end,
  stat_unit,
  emergency_name,
  publication_date
)
}
\arguments{
\item{data}{data to be passed, expects tibble}

\item{exp_end}{Expected end of the contract i.e. contract completion date}

\item{eff_start}{Effective contract signature}

\item{eff_end}{Effective end of the execution of the contract}

\item{stat_unit}{the statistical unit of measurement (can be a vector of grouping variables), i.e. variable to group by}

\item{emergency_name}{emergency name character string for which you want to evaluate the indicator, e.g. "Coronavirus" "Terremoto Aquila"}

\item{publication_date}{The date when the tender was published}
}
\value{
indicator schema as from \code{generate_indicator_schema}
}
\description{
The indicator reveals whether there has been a deviation of the contract actual execution duration from its stated/expected duration
\subsection{Motivation:}{

The red flag considers at risk companies whose contracts undergo a significant increase of their length deviation ratio - i.e., ratio between contract actual execution duration and expected duration - across the crisis
}

\subsection{Scoring Rule}{

If Test-statistic significant-> 1, otherwise -> 0
}

\subsection{Main target unit}{

This indicator targets \strong{companies} and \strong{contracting authorities}
}
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  ind_4(
    data = mock_data_core,
    publication_date = data_pubblicazione,
    exp_end = data_termine_contrattuale,
    eff_end = data_effettiva_ultimazione,
    eff_start = data_stipula_contratto,
    stat_unit = cf_amministrazione_appaltante,
    emergency_name = "coronavirus"
  )
}
}
}
\seealso{
\code{\link[lubridate]{ymd}}
\code{\link[dplyr]{filter}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{if_else}}, \code{\link[dplyr]{group_by}}, \code{\link[dplyr]{summarise}}, \code{\link[dplyr]{context}}
\code{\link[forcats]{as_factor}}
}
