% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-len-dev-across-crisis.R
\name{ind_4}
\alias{ind_4}
\title{Compute Contract Lenght deviation across the crisis indicator}
\usage{
ind_4(
  data,
  exp_end,
  eff_start,
  eff_end,
  stat_unit,
  emergency_name,
  publication_date
)
}
\arguments{
\item{data}{This argument should be a data frame or tibble containing the data you want to use to calculate the indicator.}

\item{exp_end}{Expected end of the contract i.e. contract completion date}

\item{eff_start}{This argument corresponds to the name of the column in data containing the effective start date for each contract.}

\item{eff_end}{This argument corresponds to the name of the column in data containing the effective end date for each contract.}

\item{stat_unit}{This argument should be a character string specifying the statistical unit of measurement or aggregation variable for the indicator. In this indicator both companies and contracting authorities are the targets.}

\item{emergency_name}{This argument should be a character string specifying the name of the emergency or event you are analyzing. Examples could include "Coronavirus" or "Terremoto Aquila".}

\item{publication_date}{This argument corresponds to the name of the column in data containing the publication date for each notice or report.}
}
\value{
indicator schema as from \code{generate_indicator_schema}
}
\description{
The indicator reveals whether there has been a deviation of the contract actual execution duration from its stated/expected duration
\subsection{Motivation:}{

The red flag considers at risk companies whose contracts undergo a significant increase of their length deviation ratio - i.e., ratio between contract actual execution duration and expected duration - across the crisis
}

\subsection{Scoring Rule}{

If Test-statistic significant-> 1, otherwise -> 0
}

\subsection{Main target unit}{

This indicator targets \strong{companies} and \strong{contracting authorities}
}
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  ind_4(
    data = mock_data_core,
    publication_date = data_pubblicazione,
    exp_end = data_termine_contrattuale,
    eff_end = data_effettiva_ultimazione,
    eff_start = data_stipula_contratto,
    stat_unit = cf_amministrazione_appaltante,
    emergency_name = "coronavirus"
  )
}
}
}
\seealso{
\code{\link[lubridate]{ymd}}
\code{\link[dplyr]{filter}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{if_else}}, \code{\link[dplyr]{group_by}}, \code{\link[dplyr]{summarise}}, \code{\link[dplyr]{context}}
}
