% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05bis-awarded-notice-communication.R
\name{ind_5_bis}
\alias{ind_5_bis}
\title{Compute Awarded notice communication default indicator}
\usage{
ind_5_bis(
  data,
  publ_date,
  flag_agg,
  cpv,
  outbreak_starting_date = lubridate::ymd("2017-06-30"),
  division = "33",
  stat_unit
)
}
\arguments{
\item{data}{bndcp data}

\item{publ_date}{The date when the tender was published}

\item{flag_agg}{PARAM_DESCRIPTION}

\item{cpv}{Common Procurement Vocabulary.The main vocabulary is based on a tree structure made up with codes of up to 9 digits (an 8 digit code plus a check digit). This combination of digits is associated with a wording that describes the type of supplies, works or services defining the subject of the contract.}

\item{outbreak_starting_date}{the date of the emergency outbreak, Default: lubridate::ymd("2017-06-30")}

\item{division}{initial two digits from cpv, Default: '33'}

\item{stat_unit}{statistical unit of measurement and stat_uniting}
}
\value{
indicator schema as from \code{generate_indicator_schema}
}
\description{
The indicator reveals the fraction of contracts without award notice communication.
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
\seealso{
\code{\link[lubridate]{ymd}}
\code{\link[dplyr]{mutate}}, \code{\link[dplyr]{if_else}}, \code{\link[dplyr]{stat_unit_by}}, \code{\link[dplyr]{summarise}}, \code{\link[dplyr]{distinct}}, \code{\link[dplyr]{pull}}, \code{\link[dplyr]{rename}}, \code{\link[dplyr]{mutate-joins}}, \code{\link[dplyr]{arrange}}
\code{\link[forcats]{as_factor}}
\code{\link[stringr]{str_sub}}
}
