% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05bis-awarded-notice-communication.R
\name{ind_5_bis}
\alias{ind_5_bis}
\title{Compute Awarded notice communication default indicator}
\usage{
ind_5_bis(
  data,
  publication_date,
  cpv,
  emergency_name,
  cpv_division = "33",
  stat_unit
)
}
\arguments{
\item{data}{bndcp data}

\item{publication_date}{The date when the tender was published}

\item{cpv}{Common Procurement Vocabulary.The main vocabulary is based on a tree structure made up with codes of up to 9 digits (an 8 digit code plus a check digit). This combination of digits is associated with a wording that describes the type of supplies, works or services defining the subject of the contract.}

\item{emergency_name}{emergency name character string for which you want to evaluate the indicator, e.g. "Coronavirus" "Terremoto Aquila"}

\item{cpv_division}{initial two digits from cpv, Default: '33'}

\item{stat_unit}{statistical unit of measurement and stat_uniting}
}
\value{
indicator schema as from \code{generate_indicator_schema}
}
\description{
The indicator reveals the fraction of contracts without any award notice communication.
}
\examples{
\dontrun{
if (interactive()) {
  data("mock_data_core")
  ind_5_bis(
    data = mock_data_core,
    publication_date = data_pubblicazione,
    cpv = cod_cpv,
    emergency_name = "coronavirus"
    cpv_division = "33",
    stat_unit = provincia
  )
}
}
}
\seealso{
\code{\link[lubridate]{ymd}}
\code{\link[dplyr]{mutate}}, \code{\link[dplyr]{if_else}}, \code{\link[dplyr]{stat_unit_by}}, \code{\link[dplyr]{summarise}}, \code{\link[dplyr]{distinct}}, \code{\link[dplyr]{pull}}, \code{\link[dplyr]{rename}}, \code{\link[dplyr]{mutate-joins}}, \code{\link[dplyr]{arrange}}
\code{\link[forcats]{as_factor}}
\code{\link[stringr]{str_sub}}
}
