% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-win-share-issuer-across-crisis.R
\name{ind_5}
\alias{ind_5}
\title{Compute Excess of concentration of the winners’ distribution indicator}
\usage{
ind_5(data, stat_unit, publication_date, winners, emergency_name, cpvs, ...)
}
\arguments{
\item{data}{a dataframe containing the data to use for computing the indicator.}

\item{stat_unit}{name of the variable in \code{data} containing the target unit ID (in this case, contracting authority).}

\item{publication_date}{name of the variable in \code{data} containing the publication date of each contract.}

\item{winners}{name of the variable in \code{data} containing the winning company ID of each contract.}

\item{emergency_name}{string specifying the name of the emergency to consider. Examples could include "Coronavirus" or
"Terremoto Centro Italia 2016-2017".}

\item{cpvs}{character vector of CPV divisions (first two digits of CPV code) on which \code{data} are filtered out.
Note: a panel of experts have already chosen which CPV divisions are most affected by which emergency.}

\item{...}{other parameters to pass to \code{\link[=generate_indicator_schema]{generate_indicator_schema()}}, such as \code{country_name} (default: Italy).}
}
\value{
indicator schema as from \code{\link[=generate_indicator_schema]{generate_indicator_schema()}}.
}
\description{
The indicator compares the concentration degree of the winners’ distribution of a contracting authority based
on the contracts issued after the emergency with respect to what occurred before, with reference to the contracts belonging
to the pertinent market.
\subsection{Motivation}{

The red flag considers at risk contracting authorities showing an increase in the concentration of the winners’ distribution
across the crisis, thus highlighting a high degree of awarding their contracts to an increasingly small number of firms.
}

\subsection{Scoring rule}{

For each contracting authority, a comparison of concentration index (e.g., inverse of Gini-normalised for heterogeneity) -
discretised into five categories - before and after the outbreak is performed.
If concentration category across the emergency increases, this indicator will be equal to 1. Otherwise, it will be equal to 0.
}

\subsection{Main target unit}{

This indicator targets \strong{contracting authorities}
}
}
\examples{
\dontrun{
if (interactive()) {
  mock_data_core <- mock_data_core |>
    tidyr::unnest(aggiudicatari, keep_empty = TRUE)
  ind_5(
    data = mock_data_core,
    publication_date = data_pubblicazione,
    stat_unit = cf_amministrazione_appaltante,
    winners = codice_fiscale,
    emergency_name = "coronavirus"
  )
}
}
}
\seealso{
\code{\link[lubridate]{ymd}}
\code{\link[dplyr]{filter}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{if_else}}, \code{\link[dplyr]{group_by}}, \code{\link[dplyr]{summarise}}, \code{\link[dplyr]{context}}
}
