% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-win-share-issuer-across-crisis.R
\name{ind_5}
\alias{ind_5}
\title{Compute Winner's share of issuer's contract across the crisis indicator}
\usage{
ind_5(data, stat_unit, publication_date, winners, emergency_name)
}
\arguments{
\item{data}{This argument should be a data frame or tibble containing the data you want to use to calculate the indicator.}

\item{stat_unit}{This argument should be a character string specifying the statistical unit of measurement or aggregation variable for the indicator. In this indicator both companies and contracting authorities are the targets.}

\item{publication_date}{This argument corresponds to the name of the column in data containing the publication date for each notice or report.}

\item{winners}{This argument corresponds to the name of the column in data containing the winning companies for each contract. This column should contain character or factor values.}

\item{emergency_name}{This argument should be a character string specifying the name of the emergency or event you are analyzing. Examples could include "Coronavirus" or "Terremoto Aquila".}
}
\value{
indicator schema as from \code{generate_indicator_schema}
}
\description{
The indicator reveals, for each issuer, the share of contracts awarded to the same company out of the total number of contracts issued
\subsection{Motivation:}{

The red flag considers at risk companies that \strong{exceptionally increase} \emph{their competitive power} over the emergency outbreak, as a consequence of a high proportion of contracts awarded by the same contracting authority
}

\subsection{Scoring Rule}{

If Test-statistic significant-> 1, otherwise -> 0
}

\subsection{Main target unit}{

This indicator targets \strong{companies} and \strong{contracting authorities}
}
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  ind_5(
    data = mock_data_core,
    publication_date = data_pubblicazione,
    stat_unit = cf_amministrazione_appaltante,
    winners = denominazione,
    emergency_name = "coronavirus"
  )
}
}
}
\seealso{
\code{\link[lubridate]{ymd}}
\code{\link[dplyr]{filter}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{if_else}}, \code{\link[dplyr]{group_by}}, \code{\link[dplyr]{summarise}}, \code{\link[dplyr]{context}}
}
