% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-communication-def-across-crisis.R
\name{ind_6}
\alias{ind_6}
\title{Compute Award communication default across the crisis indicator}
\usage{
ind_6(
  data,
  publication_date,
  emergency_name,
  award_col,
  stat_unit,
  test_type,
  cpvs,
  ...
)
}
\arguments{
\item{data}{a dataframe containing the data to use for computing the indicator.}

\item{publication_date}{name of the variable in \code{data} containing the publication date of each contract.}

\item{emergency_name}{string specifying the name of the emergency to consider. Examples could include "Coronavirus" or
"Terremoto Centro Italia 2016-2017".}

\item{award_col}{name of the variable in \code{data}containing the award notice ID for each contract.}

\item{stat_unit}{name of the variable in \code{data} containing the target unit ID (in this case, contracting authority).}

\item{test_type}{string specifying the statistical test to use for computing the indicator. Available options are "barnard",
"fisher", or "z-test".}

\item{cpvs}{character vector of CPV divisions (first two digits of CPV code) on which \code{data} are filtered out.
Note: a panel of experts have already chosen which CPV divisions are most affected by which emergency.}

\item{...}{other parameters to pass to \code{\link[=generate_indicator_schema]{generate_indicator_schema()}}, such as \code{country_name} (default: Italy).}
}
\value{
indicator schema as from \code{\link[=generate_indicator_schema]{generate_indicator_schema()}}.
}
\description{
The indicator detects an increase across the emergency in the rate of contracts for which the contracting authorities
have not communicated the award details to Anticorruption Authority, with reference to contracts belonging to relevant
economic market.
\subsection{Motivation}{

The red flag considers at risk those contracting authorities with an increase in the communication default rate of award
information, hence a worsening of their communication duties.
}

\subsection{Scoring rule}{

The computation procedure returns \emph{1 - p-value} of the involved test (so that high values of the indicator correspond
to high levels of corruption risk). When computing the composite, it will be dichotomised to 1 if statistical test is
significant, and 0 otherwise (see \code{\link[=normalise]{normalise()}}).
}

\subsection{Main target unit}{

This indicator targets \strong{contracting authorities}.
}
}
\examples{
\dontrun{
if (interactive()) {
  ind_6(
    data = mock_data_core,
    publication_date = data_pubblicazione,
    emergency_name = "coronavirus",
    award_col = id_aggiudicazione,
    stat_unit = cf_amministrazione_appaltante,
    test_type = "fisher"
  )
}
}
}
\seealso{
\code{\link[lubridate]{ymd}}
\code{\link[dplyr]{mutate}}, \code{\link[dplyr]{if_else}}, \code{\link[dplyr]{summarise}}, \code{\link[dplyr]{distinct}}, \code{\link[dplyr]{pull}}, \code{\link[dplyr]{rename}}, \code{\link[dplyr]{mutate-joins}}, \code{\link[dplyr]{arrange}}
\code{\link[forcats]{as_factor}}
\code{\link[stringr]{str_sub}}
}
