% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-communication-def-across-crisis.R
\name{ind_6}
\alias{ind_6}
\title{Compute Communication default across the crisis indicator}
\usage{
ind_6(data, publication_date, emergency_name, award_col, stat_unit, test_type)
}
\arguments{
\item{data}{mock_data_core example bdncp data}

\item{publication_date}{The date when the tender was published}

\item{emergency_name}{emergency name character string for which you want to evaluate the indicator, e.g. "Coronavirus" "Terremoto Aquila"}

\item{award_col}{column indentifying id for that contract award}

\item{stat_unit}{statistical unit of measurement and stat_uniting}

\item{test_type}{type of the test we would like to apply}
}
\value{
indicator schema as from \code{generate_indicator_schema}
}
\description{
The indicator reveals whether - and to what extent - the duty of contracting authorities to communicate the public procurement procedure activation to the Anticorruption Authority ended with a technical failure
\subsection{Motivation:}{

The red flag considers at risk contracting authorities \strong{who fail to accomplish their duty} to communicate the public procurement procedure activation to the Anticorruption Authority
}

\subsection{Scoring Rule}{

If Test-statistic significant-> 1, otherwise -> 0
}

\subsection{Main target unit}{

This indicator targets \strong{contracting authorities}
}
}
\examples{
\dontrun{
if (interactive()) {
  data("mock_data_core")
  ind_6(
    data = mock_data_core,
    publication_date = data_pubblicazione,
    emergency_name = "coronavirus",
    award_col = id_aggiudicazione,
    stat_unit = provincia,
    test_type = "fisher"
  )
}
}
}
\seealso{
\code{\link[lubridate]{ymd}}
\code{\link[dplyr]{mutate}}, \code{\link[dplyr]{if_else}}, \code{\link[dplyr]{stat_unit_by}}, \code{\link[dplyr]{summarise}}, \code{\link[dplyr]{distinct}}, \code{\link[dplyr]{pull}}, \code{\link[dplyr]{rename}}, \code{\link[dplyr]{mutate-joins}}, \code{\link[dplyr]{arrange}}
\code{\link[forcats]{as_factor}}
\code{\link[stringr]{str_sub}}
}
