% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-os-opportunistic-company.R
\name{ind_7}
\alias{ind_7}
\title{Compute One-shot opportunistic companies over the crisiss indicator}
\usage{
ind_7(data, final_award_date, emergency_name, stat_unit, years_before, ...)
}
\arguments{
\item{data}{This argument should be a data frame or tibble containing the data you want to use to calculate the indicator.}

\item{final_award_date}{This argument corresponds to the name of the column in data containing the date of the contract award, as recorded in the minutes or official documentation. Although the column values are stored as character strings, the function will automatically convert them to date objects when needed.}

\item{emergency_name}{This argument should be a character string specifying the name of the emergency or event you are analyzing. Examples could include "Coronavirus" or "Terremoto Aquila".}

\item{stat_unit}{This argument should be a character string specifying the statistical unit of measurement or aggregation variable for the indicator. In this indicator both companies are the targets.}

\item{years_before}{This argument specifies how many years before the contract date we need to look for the presence of awards to a specific company.}

\item{...}{other parameters to pass to \code{generate_indicator_schema} as \code{country_name} if that not Italy, which is default behavior.}
}
\description{
The indicator focuses on companies that after the emergency outbreak were awarded one or more public contracts having participated in the public procurement process without winning in the \code{years_before} the emergency outbreak
\subsection{Motivation:}{

The red flag considers at risk those companies that \strong{show a "one- shot opportunistic behaviour"}, that is, companies that after the emergency outbreak were awarded one or more public contracts but \emph{did not show any competitive power in the previous years}
}

\subsection{Scoring Rule}{

If a company wins one or more contracts after the emergency outbreak having participated in the public procurement process without winning in the 5 years before the emergency outbreak -> 1, otherwise 0
}

\subsection{Main target unit}{

This indicator targets \strong{companies}
}
}
\examples{
\dontrun{
if (interactive()) {
  data("mock_data_core")
  ind_7(
    data = mock_data_core,
    final_award_date = data_aggiudicazione_definitiva,
    stat_unit = cf_amministrazione_appaltante,
    emergency_name = "coronavirus",
    years_before = 1
  )
}
}
}
\seealso{
\code{\link[lubridate]{ymd}}, \code{\link[lubridate]{interval}}, \code{\link[lubridate]{period}}
\code{\link[tidyr]{nest}}
\code{\link[dplyr]{mutate}}, \code{\link[dplyr]{across}}, \code{\link[dplyr]{reexports}}, \code{\link[dplyr]{if_else}}, \code{\link[dplyr]{group_by}}
}
