% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-os-opportunistic-company.R
\name{ind_7}
\alias{ind_7}
\title{Compute One-shot opportunistic companies over the crisiss indicator}
\usage{
ind_7(data, award_date, emergency_name, stat_unit, years_before = 1)
}
\arguments{
\item{data}{mock_data_core exmaple data}

\item{emergency_name}{emergency name character string for which you want to evaluate the indicator, e.g. "Coronavirus" "Terremoto Aquila"}

\item{stat_unit}{The unique ID Code that identifies the awarded company (ex. VAT or Tax Number)}

\item{publication_date}{The date when the tender was published}
}
\value{
indicator schema as from \code{generate_indicator_schema}
}
\description{
The indicator reveals whether the company that won the public tender or contract has a previous history of at least one year since the opening of the tender.
}
\examples{
\dontrun{
if (interactive()) {
  data("mock_data_core")
  ind_7(
    data = mock_data_core,
    award_date = data_aggiudicazione_definitiva,
    stat_unit = cf_amministrazione_appaltante,
    emergency_name = "coronavirus"
  )
}
}
}
\seealso{
\code{\link[lubridate]{ymd}}, \code{\link[lubridate]{character(0)}}, \code{\link[lubridate]{interval}}, \code{\link[lubridate]{period}}
\code{\link[tidyr]{nest}}
\code{\link[dplyr]{mutate}}, \code{\link[dplyr]{across}}, \code{\link[dplyr]{reexports}}, \code{\link[dplyr]{if_else}}, \code{\link[dplyr]{group_by}}
\code{\link[forcats]{as_factor}}
}
