% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-os-opportunistic-company.R
\name{ind_7}
\alias{ind_7}
\title{Compute One-shot opportunistic companies over the crisiss indicator}
\usage{
ind_7(data, final_award_date, emergency_name, stat_unit, years_before)
}
\arguments{
\item{data}{mock_data_core exmaple data}

\item{final_award_date}{Date of award, as per the minutes}

\item{emergency_name}{emergency name character string for which you want to evaluate the indicator, e.g. "Coronavirus" "Terremoto Aquila"}

\item{stat_unit}{The unique ID Code that identifies the awarded company (ex. VAT or Tax Number)}

\item{years_before}{int how many years we have to}
}
\value{
indicator schema as from \code{generate_indicator_schema}
}
\description{
The indicator focuses on companies that after the emergency outbreak were awarded one or more public contracts having participated in the public procurement process without winning in the 5 years before the emergency outbreak
\subsection{Motivation:}{

The red flag considers at risk those companies that \strong{show a "one- shot opportunistic behaviour"}, that is, companies that after the emergency outbreak were awarded one or more public contracts but \emph{did not show any competitive power in the previous years}
}

\subsection{Scoring Rule}{

If a company wins one or more contracts after the emergency outbreak having participated in the public procurement process without winning in the 5 years before the emergency outbreak -> 1, otherwise 0
}

\subsection{Main target unit}{

This indicator targets \strong{companies}
}
}
\examples{
\dontrun{
if (interactive()) {
  data("mock_data_core")
  ind_7(
    data = mock_data_core,
    final_award_date = data_aggiudicazione_definitiva,
    stat_unit = cf_amministrazione_appaltante,
    emergency_name = "coronavirus",
    years_before = 1
  )
}
}
}
\seealso{
\code{\link[lubridate]{ymd}}, \code{\link[lubridate]{interval}}, \code{\link[lubridate]{period}}
\code{\link[tidyr]{nest}}
\code{\link[dplyr]{mutate}}, \code{\link[dplyr]{across}}, \code{\link[dplyr]{reexports}}, \code{\link[dplyr]{if_else}}, \code{\link[dplyr]{group_by}}
}
