% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-pre-exist-contracts-ext-after-emerg-outb.R
\name{ind_8}
\alias{ind_8}
\title{Compute Pre-existing contracts modified after the crisis indicator}
\usage{
ind_8(
  data,
  publication_date,
  stat_unit,
  variant_date,
  emergency_name,
  months_win = 6,
  cpvs,
  ...
)
}
\arguments{
\item{data}{bndcp data}

\item{publication_date}{The date when the tender was published}

\item{stat_unit}{tatistical unit of measurement, aggregation variable, the indicator target. In this case the identifier of agency or winners.}

\item{variant_date}{Date of contract variation}

\item{emergency_name}{emergency name character string for which you want to evaluate the indicator, e.g. "Coronavirus" "Terremoto Aquila"}

\item{months_win}{time window for variation to be happening}

\item{cpvs}{a vector of cpv on which contracts are filtered}

\item{...}{other parameters for generate_indicator_schema as country_name}
}
\value{
indicator schema as from \code{generate_indicator_schema}
}
\description{
The indicator reveals whether a pre- existing awarded contract has been modified after the emergency outbreak
\subsection{Motivation:}{

The red flag considers at risk contracts awarded before the emergency outbreak, but modified through variants after 6 months (parametric window) from the emergency outbreak
}

\subsection{Scoring Rule}{

If an awarded contract before the outbreak is modified through variants after 6 months from the emergency outbreak -> 1, otherwise -> 0
}

\subsection{Main target unit}{

This indicator targets \strong{companies} and \strong{contracting authorities}
}
}
\examples{
\dontrun{
if (interactive()) {
  data("mock_data_core")
  ind_8(
    data = mock_data_core,
    publication_date = data_pubblicazione,
    stat_unit = cf_amministrazione_appaltante,
    variant_date = data_approvazione_variante,
    months_win = 6,
    emergency_name = "coronavirus"
  )
}
}
}
\seealso{
\code{\link[dplyr]{mutate}}, \code{\link[dplyr]{if_else}}, \code{\link[dplyr]{group_by}}, \code{\link[dplyr]{summarise}}, \code{\link[dplyr]{context}}, \code{\link[dplyr]{n_distinct}}, \code{\link[dplyr]{rename}}
\code{\link[lubridate]{ymd}}, \code{\link[lubridate]{interval}}
}
