% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-pre-exist-contracts-ext-after-emerg-outb.R
\name{ind_8}
\alias{ind_8}
\title{Compute Pre-existing contracts modified after the crisis indicator}
\usage{
ind_8(
  data,
  publication_date,
  stat_unit,
  variant_date,
  emergency_name,
  months_win = 6
)
}
\arguments{
\item{data}{This argument should be a data frame or tibble containing the data you want to use to calculate the indicator.}

\item{publication_date}{This argument corresponds to the name of the column in data containing the publication date for each notice or report.}

\item{stat_unit}{This argument should be a character string specifying the statistical unit of measurement or aggregation variable for the indicator. In this indicator both companies and contracting authorities are the targets.}

\item{variant_date}{This argument corresponds to the name of the column in data containing the date of each contract variants.}

\item{emergency_name}{This argument should be a character string specifying the name of the emergency or event you are analyzing. Examples could include "Coronavirus" or "Terremoto Aquila".}

\item{months_win}{This argument specifies the time window for contract variation to be considered when identifying a relevant pre and post contract variation period. This value should be numeric and indicates the duration of the time window in months.}
}
\value{
indicator schema as from \code{generate_indicator_schema}
}
\description{
The indicator reveals whether a pre- existing awarded contract has been modified after the emergency outbreak
\subsection{Motivation:}{

The red flag considers at risk contracts awarded before the emergency outbreak, but modified through variants after 6 months (parametric window) from the emergency outbreak
}

\subsection{Scoring Rule}{

If an awarded contract before the outbreak is modified through variants after 6 months from the emergency outbreak -> 1, otherwise -> 0
}

\subsection{Main target unit}{

This indicator targets \strong{companies} and \strong{contracting authorities}
}
}
\examples{
\dontrun{
if (interactive()) {
  data("mock_data_core")
  ind_8(
    data = mock_data_core,
    publication_date = data_pubblicazione,
    stat_unit = cf_amministrazione_appaltante,
    variant_date = data_approvazione_variante,
    months_win = 6,
    emergency_name = "coronavirus"
  )
}
}
}
\seealso{
\code{\link[dplyr]{mutate}}, \code{\link[dplyr]{if_else}}, \code{\link[dplyr]{group_by}}, \code{\link[dplyr]{summarise}}, \code{\link[dplyr]{context}}, \code{\link[dplyr]{n_distinct}}, \code{\link[dplyr]{rename}}
\code{\link[lubridate]{ymd}}, \code{\link[lubridate]{interval}}
}
