% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-lengthy-contract.R
\name{ind_9}
\alias{ind_9}
\title{compute Lenghty Contracts indicator}
\usage{
ind_9(
  data,
  publication_date,
  stat_unit,
  cpv,
  eff_start,
  eff_end,
  emergency_name
)
}
\arguments{
\item{data}{bndcp data}

\item{publication_date}{The date when the tender was published}

\item{stat_unit}{the statistical unit of measurement (can be a vector of grouping variables), i.e. variable to group by}

\item{cpv}{Common Procurement Vocabulary. The main vocabulary is based on a tree structure made up with codes of up to 9 digits (an 8 digit code plus a check digit). This combination of digits is associated with a wording that describes the type of supplies, works or services defining the subject of the contract}

\item{eff_start}{Effective start of the execution of the contract}

\item{eff_end}{Effective end of the execution of the contract}

\item{emergency_name}{emergency name character string for which you want to evaluate the indicator, e.g. "Coronavirus" "Terremoto Aquila"}
}
\value{
indicator schema as from \code{generate_indicator_schema}
}
\description{
The indicator reveals whether the duration of the contract is significantly longer than the average length of awarded contracts after the emergency.
}
\examples{
\dontrun{
if (interactive()) {
  ind_9(
    data = mock_data_core, publication_date = data_pubblicazione,
    stat_unit = cf_amministrazione_appaltante,
    cpv = cod_cpv,
    eff_start = ...,
    eff_end = ...,
    emergency_name = ...
  )
}
}
}
\seealso{
\code{\link[dplyr]{mutate}}, \code{\link[dplyr]{if_else}}, \code{\link[dplyr]{filter}}, \code{\link[dplyr]{pull}}, \code{\link[dplyr]{group_by}}, \code{\link[dplyr]{summarise}}, \code{\link[dplyr]{context}}, \code{\link[dplyr]{rename}}
\code{\link[lubridate]{ymd}}
\code{\link[forcats]{as_factor}}
\code{\link[stringr]{str_sub}}
\code{\link[rlang]{as_string}}
}
