% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-lengthy-contract.R
\name{ind_9}
\alias{ind_9}
\title{Compute Lengthy contracts indicator}
\usage{
ind_9(
  data,
  exp_start,
  exp_end,
  eff_start,
  eff_end,
  stat_unit,
  emergency_name,
  publication_date
)
}
\arguments{
\item{data}{data to be passed, expects tibble}

\item{exp_start}{Expected initial execution of the contract}

\item{exp_end}{Expected end of the contract i.e. contract completion date}

\item{eff_start}{Effective contract signature}

\item{eff_end}{Effective end of the execution of the contract}

\item{stat_unit}{the statistical unit of measurement (can be a vector of grouping variables), i.e. variable to group by}

\item{emergency_name}{emergency name character string for which you want to evaluate the indicator, e.g. "Coronavirus" "Terremoto Aquila"}

\item{publication_date}{The date when the tender was published}
}
\value{
indicator schema as from \code{generate_indicator_schema}
}
\description{
Divergence in duration between expected and effective times of execution of the contract
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  ind_9(
    data = mock_data_core,
    publication_date = data_pubblicazione,
    exp_start = data_esecutivita_contratto
    exp_end = data_termine_contrattuale,
    eff_end = data_effettiva_ultimazione,
    eff_start = data_stipula_contratto,
    stat_unit = cf_amministrazione_appaltante,
    emergency_name = "coronavirus"
  )
  }
}
}
\seealso{
\code{\link[lubridate]{ymd}}
\code{\link[dplyr]{filter}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{if_else}}, \code{\link[dplyr]{group_by}}, \code{\link[dplyr]{summarise}}, \code{\link[dplyr]{context}}
\code{\link[forcats]{as_factor}}
}
