% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-lengthy-contract.R
\name{ind_9}
\alias{ind_9}
\title{Compute Lenghty Contracts indicator}
\usage{
ind_9(data, publication_date, stat_unit, eff_start, eff_end, emergency_name)
}
\arguments{
\item{data}{This argument should be a data frame or tibble containing the data you want to use to calculate the indicator.}

\item{publication_date}{This argument corresponds to the name of the column in data containing the publication date for each notice or report.}

\item{stat_unit}{This argument should be a character string specifying the statistical unit of measurement or aggregation variable for the indicator. In this indicator both companies and contracting authorities are the targets.}

\item{eff_start}{This argument corresponds to the name of the column in data containing the effective start date for each contract.}

\item{eff_end}{This argument corresponds to the name of the column in data containing the effective end date for each contract.}

\item{emergency_name}{This argument should be a character string specifying the name of the emergency or event you are analyzing. Examples could include "Coronavirus" or "Terremoto Aquila".}
}
\value{
indicator schema as from \code{generate_indicator_schema}
}
\description{
The indicator reveals whether the duration of a contract is significantly longer than the average length of awarded contracts after the emergency outbreak
\subsection{Motivation:}{

The red flag considers at risk contracts won after the emergency outbreak whose duration is not justified by the nature of the crisis, that is, contracts longer than the average duration of contracts won after the emergency outbreak
}

\subsection{Scoring Rule}{

The output will give $1 - pvalue$, which will then be dichotomised to 1 if statistical test is significant, 0 otherwise.
}

\subsection{Main target unit}{

This indicator targets \strong{companies} and \strong{contracting authorities}
}
}
\examples{
\dontrun{
if (interactive()) {
  ind_9(
    data = mock_data_core,
    publication_date = data_pubblicazione,
    stat_unit = cf_amministrazione_appaltante,
    eff_start = data_inizio_effettiva,
    eff_end = data_effettiva_ultimazione,
    emergency_name = "coronavirus"
  )
}
}
}
\seealso{
\code{\link[dplyr]{mutate}}, \code{\link[dplyr]{if_else}}, \code{\link[dplyr]{filter}}, \code{\link[dplyr]{pull}}, \code{\link[dplyr]{group_by}}, \code{\link[dplyr]{summarise}}, \code{\link[dplyr]{context}}, \code{\link[dplyr]{rename}}
\code{\link[lubridate]{ymd}}
\code{\link[forcats]{as_factor}}
\code{\link[stringr]{str_sub}}
\code{\link[rlang]{as_string}}
}
