% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mock_data_core}
\alias{mock_data_core}
\title{BDNCP open data from ANAC}
\format{
\subsection{\code{who}}{

A data frame with 7,240 rows and 60 columns:
\describe{
\item{cig, cig_accordo_quadro}{alphanumeric code generated by the ANAC SIMOG system}
\item{numero_gara, oggetto_gara}{bidding related info}
\item{importo_complessivo_gara}{bidding amount}
\item{year}{Year}
\item{year}{Year}
\item{year}{Year}
\item{year}{Year}
\item{year}{Year}
\item{year}{Year}
\item{year}{Year}
\item{year}{Year}
\item{year}{Year}
\item{year}{Year}
\item{year}{Year}
\item{year}{Year}
\item{year}{Year}
\item{year}{Year}
\item{year}{Year}
\item{year}{Year}
\item{year}{Year}
\item{year}{Year}
\item{year}{Year}
\item{year}{Year}
\item{year}{Year}
\item{year}{Year}
...
}
}
}
\source{
\url{https://dati.anticorruzione.it/opendata}
}
\usage{
mock_data_core
}
\description{
available data can be consulted and downloaded through the
appropriate subsections; in particular, in the "Analytics" section
there is a dashboard for browsing and analyzing data on public contracts,
while in the "Dataset" section a series of freely downloadable
JSON and CSV and XML files are available.
}
\keyword{datasets}
